Use int instead of int32/int64.
Use int instead of int32 for page numbers. Page numbers are always going to be well under the 32 bit limit, so make the API simpler by making them ints.

Use int instead of int64 for measurements in seconds. The time values are never going to be above the 32 bit limit, and using an int makes the API simpler.